pyproject.toml
[project]
name = "chirpstack-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chirpstack-api>=4.13.0",
    "paho-mqtt>=2.1.0",
    "rich>=14.1.0",
    "requests>=2.31.0",
    "typer>=0.16.0",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "fastapi[all]>=0.116.1",
    "sqlmodel>=0.0.24",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
]
